refactor: reorganize and update imports for `LoginSessionCommand` components
- Relocated `ValueObjects` and `DTO` files to appropriate `Application` and `Domain` subdirectories.
- Updated import paths in relevant validators, handlers, and repository interfaces.
- Replaced `VerificationService` with `UserService` in `index.ts`.
