+++ user-service/src/index.ts
@@ -15,7 +15,7 @@ import { GetUserService } from './Application/Services/Concrete/GetUserService.j
 import { NotificationError } from './Shared/Errors/NotificationError.js';
 import { LoginUserService } from './Application/Services/Concrete/LoginUserService.js';
 import { LogoutUserService } from './Application/Services/Concrete/LogoutUserService.js';
-import {VerificationService} from "./Application/Services/Concrete/VerificationService.js";
+import {UserService} from "./Application/Services/Concrete/UserService.js";
 
 declare module 'fastify' {
   export interface FastifyInstance {
@@ -48,7 +48,7 @@ async function main() {
   const editUserService = new EditUserService(userRepository, notificationError);
   const deleteUserService = new DeleteUserService(userRepository, notificationError);
   const getUserService = new GetUserService(userRepository, notificationError);
-  const verificationService = new VerificationService(userRepository, notificationError);
+  const verificationService = new UserService(userRepository, notificationError);
   const loginUserService = new LoginUserService(userRepository, notificationError);
   const logoutUserService = new LogoutUserService(userRepository, notificationError);
